refactor(#727): move the remaining three strategy implementations + library-miss probe into lookup/strategies/
Pure code motion, PR 4b of the orchestrator decomposition (#722) — the biggest rewire batch of the program. search_compilations_for_track (443 lines, moved INTACT per the explicit non-goal), its _ProbeResult probe-union alias, and its private telemetry helper _log_album_title_fallback land in the existing track_on_compilation.py shell; search_song_as_artist lands in the song_as_artist.py shell; search_library_with_fallback (including the #718 wrong-artist/VA-comp backstop) lands in the artist_plus_album.py shell; and the #583 step-3a library-miss Discogs probe _library_miss_discogs_search gets its own strategy-adjacent module lookup/strategies/library_miss.py (not a pipeline strategy — perform_lookup calls it directly, so the module docstring says exactly that). The orchestrator keeps consumer imports for all four (build_strategies partials + the step-3a call) and drops the seventeen imports whose last consumer moved; no strategies module imports lookup.orchestrator, preserving the downward dependency direction. Only additions beyond the relocated bytes are the module docstring, import blocks, and logger lines the program mechanics allow.
